Added Hypercubic Polynomial Quadratic Sieve
This algorithm is the record holder for numbers less than about ~100 decimal
digits.  It is now the default method which cl-factoring:factor uses.  There are
some tunable parameters, but I don't understand it enough to know how to use
them, but it seems to work well as a dumb black box.

This is a near verbatim copy of the code by Uli Meyer (some slight modifications
to interface with normal Common Lisp style and the library).