Switch to Pollard Rho as a bug free implementation
This is a work-around for issue #1 on the Github project page, that the
algorithm has an error rather than factoring the number 327146098439579.
While I gain the knowledge necessary to fix bug in HMPQS, I am switching
the factoring method to a simpler method.

After checking what effect this will have, I've found that it is
significantly slower on average than the HMPQS method, even in number
ranges where it should win.  That isn't too surprising as almost zero
effort has been put into optimizing it.